Move delayed neutron frac input logic to neutronics plugin.
